Refactor to use Context API for network status
