+++ src/components/ModelingSidebar/ModelingPanes/KclEditorPane.tsx
@@ -33,6 +33,9 @@ import {
   syntaxHighlighting,
   defaultHighlightStyle,
 } from '@codemirror/language'
+import { useModelingContext } from 'hooks/useModelingContext'
+import { useNetworkContext } from 'hooks/useNetworkContext'
+import { NetworkHealthState } from 'hooks/useNetworkStatus'
 import interact from '@replit/codemirror-interact'
 import { kclManager, editorManager, codeManager } from 'lib/singletons'
 import { useHotkeys } from 'react-hotkeys-hook'
@@ -63,6 +66,9 @@ export const KclEditorPane = () => {
       ? getSystemTheme()
       : context.app.theme.current
   const { copilotLSP, kclLSP } = useLspContext()
+  const { overallState } = useNetworkContext()
+  const isNetworkOkay = overallState === NetworkHealthState.Ok
+  const navigate = useNavigate()
 
   useEffect(() => {
     if (typeof window === 'undefined') return
