+++ src/hooks/useNetworkContext.tsx
@@ -1,0 +1,6 @@
+import { createContext, useContext } from 'react'
+
+export const NetworkContext = createContext(null)
+export const useNetworkContext = () => {
+  return useContext(NetworkContext)
+}
